refactor!: move adapter kits to kineokit
